+++ eds-user-20210308/core/Models/DescribeUsersShrinkRequest.cs
@@ -37,6 +37,10 @@ namespace AlibabaCloud.SDK.Eds_user20210308.Models
         [Validation(Required=false)]
         public string Filter { get; set; }
 
+        [NameInMap("FilterWithAssignedResource")]
+        [Validation(Required=false)]
+        public string FilterWithAssignedResourceShrink { get; set; }
+
         [NameInMap("FilterWithAssignedResources")]
         [Validation(Required=false)]
         public string FilterWithAssignedResourcesShrink { get; set; }